requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
tensorflow==2.20.0
numpy==2.4.0
Pillow==10.1.0
python-multipart